feat: support network based streamdecks
