+++ packages/input-manager/package.json
@@ -91,7 +91,8 @@
 		"production"
 	],
 	"dependencies": {
-		"@elgato-stream-deck/node": "7.1.2",
+		"@elgato-stream-deck/node": "7.2.0",
+		"@elgato-stream-deck/tcp": "7.2.0",
 		"@julusian/jpeg-turbo": "^2.1.0",
 		"easymidi": "^3.1.0",
 		"eventemitter3": "5.0.1",
